add: navigation DanhSachBaiViet to ChiTietBaiViet (state)
