+++ src/App.js
@@ -5,6 +5,7 @@ import Home from "./components/Home";
 import Podcast from "./components/Podcast";
 import Blog from "./components/Blog";
 import FabHoiDap from "./components/FabHoiDap";
+import ChiTietBaiViet from "./components/Blog/ChiTietBaiViet";
 
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import { MainProvider } from "./contexts/MainContext";
@@ -19,6 +20,7 @@ function App() {
             <Route path="/" element={<Home />} />
             <Route path="/podcast" element={<Podcast />} />
             <Route path="/blog" element={<Blog />} />
+            <Route path="/blog/:slug" element={<ChiTietBaiViet />} />
           </Routes>
           <FabHoiDap />
         </BrowserRouter>
