+++ src/components/Blog/ChiTietBaiViet/index.jsx
@@ -1,28 +1,32 @@
 import * as React from "react";
 import { Stack } from "@mui/material";
-import FacebookIcon from '@mui/icons-material/Facebook';
-import LinkedInIcon from '@mui/icons-material/LinkedIn';
-import YouTubeIcon from '@mui/icons-material/YouTube';
-import InstagramIcon from '@mui/icons-material/Instagram';
-import Comment from '../Components/Comment';
-import Content from "../Components/Content";
-import { useMediaQueryValues } from "../../../contexts/MediaQueryContext";
+import FacebookIcon from "@mui/icons-material/Facebook";
+import LinkedInIcon from "@mui/icons-material/LinkedIn";
+import YouTubeIcon from "@mui/icons-material/YouTube";
+import InstagramIcon from "@mui/icons-material/Instagram";
+import BinhLuan from "../Components/BinhLuan";
+import NoiDungBaiViet from "../Components/NoiDungBaiViet";
+import { useMainValues } from "../../../contexts/MainContext";
 import TimKiem from "../Components/TimKiem";
 import GioiThieu from "../Components/GioiThieu";
-import BaiVietNoiBac from "../Components/BaiVietNoiBac";
+import BaiVietNoiBat from "../Components/BaiVietNoiBat";
 import DanhMuc from "../Components/DanhMuc";
+import HoatDong from "../Components/HoatDong";
+import { useLocation } from "react-router-dom";
 
 function ChiTietBaiViet() {
-  const { isMediumScreen } = useMediaQueryValues();
+  let { state } = useLocation();
+  const { isMediumScreen } = useMainValues();
   const [dagtaGioiThieu, setDaGtaGioiThieu] = React.useState({});
-  const [dataBaiVietNoiBac, setDataBaiVietNoiBac] = React.useState([]);
-  
+  const [dataBaiVietNoiBat, setDataBaiVietNoiBat] = React.useState([]);
   React.useEffect(() => {
     setDaGtaGioiThieu({
-      image: "https://cdnmedia.baotintuc.vn/Upload/4l8oGGp94lA5r6lHXppg/files/2022/03/f0conennammaylanh.jpg",
-      content: "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nulla vitae laoreet justo. Donec euismod, nisl eget ultricies aliquam, velit odio lacinia diam, in malesuada odio odio in velit. Sed non ornare nisl",
+      image:
+        "https://cdnmedia.baotintuc.vn/Upload/4l8oGGp94lA5r6lHXppg/files/2022/03/f0conennammaylanh.jpg",
+      content:
+        "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nulla vitae laoreet justo. Donec euismod, nisl eget ultricies aliquam, velit odio lacinia diam, in malesuada odio odio in velit. Sed non ornare nisl",
     });
-    setDataBaiVietNoiBac([  
+    setDataBaiVietNoiBat([
       {
         _id: 1,
         title: "Bài viết nổi bật 1",
@@ -30,7 +34,7 @@ function ChiTietBaiViet() {
       {
         _id: 2,
         title: "Bài viết nổi bật 2",
-      }
+      },
     ]);
   }, []);
 
@@ -43,23 +47,23 @@ function ChiTietBaiViet() {
         backgroundColor={"#fff"}
       >
         <Stack
-          width={isMediumScreen ? "100%" : "80%"}
+          width={isMediumScreen ? "100%" : "60%"}
           direction={isMediumScreen ? "column" : "row"}
           justifyContent={"center"}
           alignItems={"flex-start"}
         >
           <Stack
-            width={isMediumScreen ? "100%" : "70%"}
+            width={isMediumScreen ? "100%" : "60%"}
             direction={"column"}
             justifyContent={"center"}
             alignItems={"center"}
           >
-            <Content />
-            <Comment />
+            <NoiDungBaiViet content={state.chiTietBaiViet} />
+            <BinhLuan />
           </Stack>
 
           <Stack
-            width={isMediumScreen ? "100%" : "30%"}
+            width={isMediumScreen ? "100%" : "40%"}
             direction={"column"}
             justifyContent={"center"}
             alignItems={"center"}
@@ -72,23 +76,10 @@ function ChiTietBaiViet() {
               width={"100%"}
               marginBottom={2}
             >
-              <FacebookIcon 
-                fontSize="large"
-                onClick={() => {}}
-              />
-              <LinkedInIcon 
-                fontSize="large"
-                onClick={() => {}}
-              />
-              <YouTubeIcon 
-                fontSize="large"
-                onClick={() => {}}
-              />
-              <InstagramIcon 
-                fontSize="large"
-                onClick={() => {}}
-              />
-
+              <FacebookIcon fontSize="large" onClick={() => {}} />
+              <LinkedInIcon fontSize="large" onClick={() => {}} />
+              <YouTubeIcon fontSize="large" onClick={() => {}} />
+              <InstagramIcon fontSize="large" onClick={() => {}} />
             </Stack>
 
             <TimKiem
@@ -99,20 +90,16 @@ function ChiTietBaiViet() {
               ariaLabel={"toggle password visibility"}
               position={"end"}
             />
-            <GioiThieu 
-              data={dagtaGioiThieu}
-            />
+            <GioiThieu data={dagtaGioiThieu} />
 
-            <BaiVietNoiBac
-              data={dataBaiVietNoiBac}
-            />
+            <BaiVietNoiBat data={dataBaiVietNoiBat} />
 
             <DanhMuc />
+
+            <HoatDong />
           </Stack>
         </Stack>
       </Stack>
-
-
     </div>
   );
 }
