+++ src/components/Blog/Components/ItemDanhSachBaiViet/index.jsx
@@ -1,12 +1,12 @@
 import * as React from "react";
 import CssBaseline from "@mui/material/CssBaseline";
-import { Button, Skeleton, Stack, Typography } from "@mui/material";
+import { Button, Stack, Typography } from "@mui/material";
 import { formatDateTime } from "../../../../utils/common";
+import { Link } from "react-router-dom";
 
-function ItemBaiViet(props) {
-  const [dataItem, setDataItem] = React.useState(props.dataItem);
+function ItemDanhSachBaiViet(props) {
+  const [dataItem] = React.useState(props.dataItem);
   const MAX_LENGTH_CONTENT = 443;
-
   return (
     <React.Fragment>
       <CssBaseline />
@@ -66,7 +66,6 @@ function ItemBaiViet(props) {
           >
             <Typography
               fontSize={12}
-              fontSize={12}
               fontFamily={"Montserrat"}
               fontWeight={"Regular"}
             >
@@ -105,7 +104,6 @@ function ItemBaiViet(props) {
           <img
             src={dataItem && dataItem.cover_image ? dataItem.cover_image : ""}
             alt="img"
-            //responsive style maxHeight="200px"
             width={"100%"}
             height={"200px"}
             style={{
@@ -123,7 +121,6 @@ function ItemBaiViet(props) {
             color={"#333"}
             width={"100%"}
           >
-            {/* limit paraphap  length 200 ... */}
             {dataItem && dataItem.content
               ? dataItem.content.length > MAX_LENGTH_CONTENT
                 ? dataItem.content.substr(0, MAX_LENGTH_CONTENT) + "..."
@@ -131,25 +128,32 @@ function ItemBaiViet(props) {
               : ""}
           </Typography>
 
-          <Button
-            sx={{
-              border: "0.1px solid" + "#A9A9AC",
-              color: "#000",
-              borderRadius: "8px",
-              textTransform: "capitalize",
-              fontSize: "16px",
-              fontFamily: "Montserrat",
-              fontWeight: "Regular",
-              width: "100%",
-            }}
-            variant="outlined"
-          >
-            Đọc thêm
-          </Button>
+          {dataItem && dataItem.slug && (
+            <Link
+              to={`/blog/${dataItem.slug}`}
+              state={{ chiTietBaiViet: dataItem }}
+            >
+              <Button
+                sx={{
+                  border: "0.1px solid" + "#A9A9AC",
+                  color: "#000",
+                  borderRadius: "8px",
+                  textTransform: "capitalize",
+                  fontSize: "16px",
+                  fontFamily: "Montserrat",
+                  fontWeight: "Regular",
+                  width: "100%",
+                }}
+                variant="outlined"
+              >
+                Đọc thêm
+              </Button>
+            </Link>
+          )}
         </Stack>
       </Stack>
     </React.Fragment>
   );
 }
 
-export default ItemBaiViet;
+export default ItemDanhSachBaiViet;
